#' Variable map for engineering features specific to SISMA CPN dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_cpn_map) #' #' @format A list of the variables included to engineer dataset features #'
#'
data_sisma_smi_cpn_map

indicator

#'
data_sisma_smi_cpn_map

indicator_new

#'
data_sisma_smi_cpn_map

age

#'
data_sisma_smi_cpn_map

age_coarse

#'
data_sisma_smi_cpn_map

disaggregate

#'
data_sisma_smi_cpn_map

source

#'
"data_sisma_smi_cpn_map"

#' Variable map for engineering features specific to SISMA ATS Results dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_results_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_results_map

indicator

#'
data_sisma_ats_results_map

indicator_new

#'
data_sisma_ats_results_map

sex

#'
data_sisma_ats_results_map

age

#'
data_sisma_ats_results_map

age_coarse

#'
data_sisma_ats_results_map

disaggregate

#'
data_sisma_ats_results_map

result_status

#'

"data_sisma_ats_results_map"

Usage

data(data_sisma_sitelist)

Format

A list of the variables included in the site list

data_sisma_sitelist

sisma_uid

data_sisma_sitelist

provincia

data_sisma_sitelist

distrito

data_sisma_sitelist

us

Details

#' Variable map for engineering features specific to SISMA ATS Case Index/Linkage dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_ci_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_ci_map

indicator

#'
data_sisma_ats_ci_map

indicator_new

#'
data_sisma_ats_ci_map

disaggregate

#'
data_sisma_ats_ci_map

sub_group

#'
data_sisma_ats_ci_map

result_status

#'
data_sisma_ats_ci_map

age_coarse

#'

"data_sisma_ats_ci_map"

#' Variable map for engineering features specific to SISMA ATS Historical/Key Pop dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_hist_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_hist_map

indicator

#'
data_sisma_ats_hist_map

indicator_new

#'
data_sisma_ats_hist_map

disaggregate

#'
data_sisma_ats_hist_map

sub_group

#'
data_sisma_ats_hist_map

result_status

#'

"data_sisma_ats_hist_map"

#' Variable map for engineering features specific to SISMA ATS SMI dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_smi_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_smi_map

indicator

#'
data_sisma_ats_smi_map

indicator_new

#'
data_sisma_ats_smi_map

modality

#'
data_sisma_ats_smi_map

sub_group

#'
data_sisma_ats_smi_map

result_status

#'
data_sisma_ats_smi_map

age_coarse

#'
data_sisma_ats_smi_map

sex

#'

"data_sisma_ats_smi_map"

#' Variable map for engineering features specific to SISMA ATS CCS and CCD dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_ccsd_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_ccsd_map

indicator

#'
data_sisma_ats_ccsd_map

indicator_new

#'
data_sisma_ats_ccsd_map

source

#'
data_sisma_ats_ccsd_map

disaggregate

#'
data_sisma_ats_ccsd_map

disaggregate_sub

#'
data_sisma_ats_ccsd_map

result_status

#'
data_sisma_ats_ccsd_map

age_coarse

#'
data_sisma_ats_ccsd_map

sex

#'

"data_sisma_ats_ccsd_map"

#' Variable map for engineering features specific to SISMA ATS Self-testing dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_autoteste_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_autoteste_map

indicator

#'
data_sisma_ats_autoteste_map

indicator_new

#'
data_sisma_ats_autoteste_map

age

#'
data_sisma_ats_autoteste_map

age_coarse

#'
data_sisma_ats_autoteste_map

sex

#'
data_sisma_ats_autoteste_map

disaggregate

#'
data_sisma_ats_autoteste_map

source

#'

"data_sisma_ats_autoteste_map"

#' Variable map for engineering features specific to SISMA ATS SAAJ dataframe #' #' List of indicators #' #' @usage data(data_sisma_ats_saaj_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_ats_saaj_map

indicator

#'
data_sisma_ats_saaj_map

indicator_new

#'
data_sisma_ats_saaj_map

source

#'
data_sisma_ats_saaj_map

age

#'
data_sisma_ats_saaj_map

disaggregate

#'
data_sisma_ats_saaj_map

sub_group

#'
data_sisma_ats_saaj_map

result_status

#'
data_sisma_ats_saaj_map

age_coarse

#'
data_sisma_ats_saaj_map

sex

#'

"data_sisma_ats_saaj_map"

#' Variable map for engineering features specific to SISMA TARV dataframe #' #' List of indicators #' #' @usage data(data_sisma_hiv_tarv_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_hiv_tarv_map

indicator

#'
data_sisma_hiv_tarv_map

indicator_new

#'
data_sisma_hiv_tarv_map

age

#'
data_sisma_hiv_tarv_map

age_coarse

#'
data_sisma_hiv_tarv_map

sex

#'
data_sisma_hiv_tarv_map

disaggregate

#'
data_sisma_hiv_tarv_map

source

#'

"data_sisma_hiv_tarv_map"

#' Variable map for engineering features specific to SISMA SMI Maternity dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_mat_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_mat_map

indicator

#'
data_sisma_smi_mat_map

indicator_new

#'
data_sisma_smi_mat_map

disaggregate

#'
data_sisma_smi_mat_map

disaggregate_sub

#'
data_sisma_smi_mat_map

source

#'

"data_sisma_smi_mat_map"

#' Variable map for engineering features specific to SISMA CCR Maternity dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_ccr_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_ccr_map

indicator

#'
data_sisma_smi_ccr_map

indicator_new

#'
data_sisma_smi_ccr_map

disaggregate

#'
data_sisma_smi_ccr_map

source

#'

"data_sisma_smi_ccr_map"

#' Variable map for engineering features specific to SISMA PrEP dataframe #' #' List of indicators #' #' @usage data(data_sisma_hiv_prep_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_hiv_prep_map

indicator

#'
data_sisma_hiv_prep_map

indicator_new

#'
data_sisma_hiv_prep_map

age

#'
data_sisma_hiv_prep_map

age_coarse

#'
data_sisma_hiv_prep_map

sex

#'
data_sisma_hiv_prep_map

disaggregate

#'
data_sisma_hiv_prep_map

source

#'

"data_sisma_hiv_prep_map"

#' Variable map for engineering features specific to SISMA APSS dataframe #' #' List of indicators #' #' @usage data(data_sisma_hiv_apss_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_hiv_apss_map

indicator

#'
data_sisma_hiv_apss_map

indicator_new

#'
data_sisma_hiv_apss_map

age

#'
data_sisma_hiv_apss_map

age_coarse

#'
data_sisma_hiv_apss_map

sex

#'
data_sisma_hiv_apss_map

source

#'

"data_sisma_hiv_apss_map"

#' Variable map for engineering features specific to SISMA ITS dataframe #' #' List of indicators #' #' @usage data(data_sisma_hiv_its_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_hiv_its_map

indicator

#'
data_sisma_hiv_its_map

indicator_new

#'
data_sisma_hiv_its_map

age

#'
data_sisma_hiv_its_map

age_coarse

#'
data_sisma_hiv_its_map

sex

#'
data_sisma_hiv_its_map

sub_group

#'
data_sisma_hiv_its_map

result_status

#'
data_sisma_hiv_its_map

disaggregate

#'
data_sisma_hiv_its_map

source

#'

"data_sisma_hiv_its_map"

#' Variable map for engineering features specific to SISMA AJM/HC/MM dataframe #' #' List of indicators #' #' @usage data(data_sisma_hiv_ajmhcmm_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_hiv_ajmhcmm_map

indicator

#'
data_sisma_hiv_ajmhcmm_map

indicator_new

#'
data_sisma_hiv_ajmhcmm_map

age

#'
data_sisma_hiv_ajmhcmm_map

sex

#'
data_sisma_hiv_ajmhcmm_map

result_status

#'
data_sisma_hiv_ajmhcmm_map

disaggregate

#'
data_sisma_hiv_ajmhcmm_map

source

#'

"data_sisma_hiv_ajmhcmm_map"

#' Variable map for engineering features specific to SISMA Advanced HIV Disease dataframe #' #' List of indicators #' #' @usage data(data_sisma_hiv_dah_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_hiv_dah_map

indicator

#'
data_sisma_hiv_dah_map

indicator_new

#'
data_sisma_hiv_dah_map

sub_group

#'
data_sisma_hiv_dah_map

disaggregate

#'
data_sisma_hiv_dah_map

sex

#'
data_sisma_hiv_dah_map

age

#'
data_sisma_hiv_dah_map

age_coarse

#'
data_sisma_hiv_dah_map

source

#'

"data_sisma_hiv_dah_map"

#' Variable map for engineering features specific to SISMA SMI Post-Partum Care dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_cpp_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_cpp_map

indicator

#'
data_sisma_smi_cpp_map

indicator_new

#'
data_sisma_smi_cpp_map

sex

#'
data_sisma_smi_cpp_map

age

#'
data_sisma_smi_cpp_map

sub_group

#'
data_sisma_smi_cpp_map

age_coarse

#'
data_sisma_smi_cpp_map

result_status

#'
data_sisma_smi_cpp_map

disaggregate

#'
data_sisma_smi_cpp_map

disaggregate_sub

#'
data_sisma_smi_cpp_map

source

#'

"data_sisma_smi_cpp_map"

#' Variable map for engineering features specific to SISMA Vaccination dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_pav_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_pav_map

indicator

#'
data_sisma_smi_pav_map

indicator_new

#'
data_sisma_smi_pav_map

sex

#'
data_sisma_smi_pav_map

age

#'
data_sisma_smi_pav_map

age_coarse

#'
data_sisma_smi_pav_map

disaggregate

#'
data_sisma_smi_pav_map

source

#'

"data_sisma_smi_pav_map"

#' Variable map for engineering features specific to SISMA SMI Sick-Child Care dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_ccd_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_ccd_map

indicator

#'
data_sisma_smi_ccd_map

indicator_new

#'
data_sisma_smi_ccd_map

sex

#'
data_sisma_smi_ccd_map

age

#'
data_sisma_smi_ccd_map

age_coarse

#'
data_sisma_smi_ccd_map

disaggregate

#'
data_sisma_smi_ccd_map

disaggregate_sub

#'
data_sisma_smi_ccd_map

result_status

#'
data_sisma_smi_ccd_map

source

#'

"data_sisma_smi_ccd_map"

#' Variable map for engineering features specific to SISMA SMI Well-Child Care dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_ccs_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_ccd_map

indicator

#'
data_sisma_smi_ccd_map

indicator_new

#'
data_sisma_smi_ccd_map

sex

#'
data_sisma_smi_ccd_map

age

#'
data_sisma_smi_ccd_map

age_coarse

#'
data_sisma_smi_ccd_map

disaggregate

#'
data_sisma_smi_ccd_map

disaggregate_sub

#'
data_sisma_smi_ccd_map

result_status

#'
data_sisma_smi_ccd_map

source

#'

"data_sisma_smi_ccs_map"

#' Variable map for engineering features specific to SISMA SMI Gynecology dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_ug_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_ccd_map

indicator

#'
data_sisma_smi_ccd_map

indicator_new

#'
data_sisma_smi_ccd_map

sex

#'
data_sisma_smi_ccd_map

age

#'
data_sisma_smi_ccd_map

age_coarse

#'
data_sisma_smi_ccd_map

disaggregate

#'
data_sisma_smi_ccd_map

disaggregate_sub

#'
data_sisma_smi_ccd_map

result_status

#'
data_sisma_smi_ccd_map

source

#'

"data_sisma_smi_ug_map"

#' Variable map for engineering features specific to SISMA SMI PF dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_pf_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_pf_map

indicator

#'
data_sisma_smi_pf_map

indicator_new

#'
data_sisma_smi_pf_map

sex

#'
data_sisma_smi_pf_map

age

#'
data_sisma_smi_pf_map

age_coarse

#'
data_sisma_smi_pf_map

disaggregate

#'
data_sisma_smi_pf_map

disaggregate_sub

#'
data_sisma_smi_pf_map

source

#'

"data_sisma_smi_pf_map"

#' Variable map for engineering features specific to SISMA SMI PF Integrated, SAAJ, & APE dataframe #' #' List of indicators #' #' @usage data(data_sisma_smi_pf_int_saaj_ape_map) #' #' @format A list of the variables included to engineer dataset features #'

#'
data_sisma_smi_pf_int_saaj_ape_map

indicator

#'
data_sisma_smi_pf_int_saaj_ape_map

indicator_new

#'
data_sisma_smi_pf_int_saaj_ape_map

source

#'
data_sisma_smi_pf_int_saaj_ape_map

disaggregate

#'
data_sisma_smi_pf_int_saaj_ape_map

disaggregate_sub

#'
data_sisma_smi_pf_int_saaj_ape_map

sub_group

#'
data_sisma_smi_pf_int_saaj_ape_map

sex

#'
data_sisma_smi_pf_int_saaj_ape_map

age_coarse

#'
data_sisma_smi_pf_int_saaj_ape_map

age

#'
data_sisma_smi_pf_int_saaj_ape_map

result_status

#'

"data_sisma_smi_pf_int_saaj_ape_map" Deduplicated SISMA site list

List of indicators