Processar: Projecções demográficas distritais do INE

process_pop_ine usa um vector de ficheiros contendo projecções demográficas distritais do INE processando eles num único ficheiro arrumado

Usage

process_pop_ine(
  file_inventory,
  input_sheets,
  age_level = "Exact",
  output_type = "MISAU"
)

Arguments

file_inventory

vector de ficheiros populacionais do INE em formato .xlsx

input_sheets

vector (em caracteres) dos anos do ficheiro para inclusão no processamento

age_level

Nível de pormenor da idade no quadro de dados arrumado (“Exato” por defeito com opção para “Agrupado” para faixas etárias de 5 anos)

output_type

formato do provincia (por defeito "MISAU")

Value

process_pop_ine devolve um quadro de dados arrumado com 9 colunas

Examples

 if (FALSE) { # \dontrun{

 df <- process_pop_ine()} # }