sismar: Transform tabular export files from SISMA

Provides a set of functions for creating analytical datasets from SISMA CSV downloads. This includes functions that reshape files into long format, remove unnecessary variables, and engineer features useful in analysis.

See also

Author

Maintainer: Joe Lara jlara@usaid.gov

Authors:

Emilton Cumbana e.cumbana@outlook.com

Karishma Srikanth ksrikanth@usaid.gov

Kirstin Lyon kirstin.c.lyon@gmail.com